+++ examples/email_validate_example.php
@@ -1,0 +1,3 @@
+<?php
+
+echo $sanitizer->validateEmail("fake.email@gmail.com") ? "true" : "false";