+++ examples/preg_replace_example.php
@@ -1,0 +1,3 @@
+<?php
+
+echo $sanitizer->usePregReplace("/([A-Z])\w+/", "This is a Text");