+++ examples/str_replace_example.php
@@ -1,0 +1,3 @@
+<?php
+
+echo $sanitizer->useStrReplace("text", "", "this is a text");